+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Stable versions are 2.xx still supported.
 * remove `EnumFull::enum_descriptor()`, does not provide much value over `enum_descriptor_static`
 * rename `EnumFull::enum_descriptor_static` -> `enum_descriptor`
 * fix `EnumFull::descriptor` for enums with `allow_alias`
+* rename `MessageDescriptor::enums` to `nested_enums`
 * add `EnumDescriptor::enclosing_message()`
 
 ## [3.0.0-alpha.9] - 2022-04-04